help: Increase read buffer size from 1100 to 2048
This fixes "Error opening helpfile." that was introduced by the rawreply
commit, which made the help string for "help set commands" larger than
1100 bytes, making the strstr(s, "\n%\n") condition fail

I have absolutely no idea why it was 1100 to begin with. It's been like
that since the first revision. 2048 is arbitrary too, but it's nicer.

Needless to say, the help.c code needs to be reviewed more deeply, at
least. Could also use that opportunity to get rid of the xmlto/xslt crap.
Not going to do that right now, i don't want to add more bugs.