msn: Update FindMembership payload to fix incorrect pending requests
It was requesting memberships for the following services:

    Messenger, Invitation, SocialNetwork, Space, Profile

Now it only requests the same services as WLM 2012:

    Messenger, IMAvailability

Fixes trac ticket 1191.

I'm going to apply this to the master branch just for the sake of
marking it as fixed, but I can't log in at all with that branch at the
moment. Yes this is silly.