Make my fork of parson non-JSON-compliant. I don't care that JavaScript's tiny little brain doesn't comprehend integers and wants everything to be a float, the adults need integers now.  Big integers. So add them, just like the old js lib did. This also lets me clean up some horrible things in the RPC module.
Try to have behaviour similar to the old lib: Throw as much as possible into
an integer, only switch to double when there's a non-int or a number larger
than what fits a "signed long long".

Yes this is horrible, and no I don't think the parson upstreams will accept
this. But the alternative is to stay compliant to an accidentally popular
data interchange format based on accidentally popular joke language and do
a lot of manual string-to-int64 parsing in the Twitter codebase. How about
NOPE.