Add python script to generate help.txt, replacing xslt stuff
This would mean changing one build dependency for another, but one that
is way more common at least. (Fun fact: the xslt stuff depends on perl)

It generates *almost* the same thing as the xslt - a bit better if you
ask me, since it correctly handles a few <emphasis> tags in the middle
of the text, which were previously stripped. One example of that is:

    Favo<emphasis>u</emphasis>rite the given user [...]

Outputs "Favo\x02u\x02rite" with this script, "Favorite" with the xslt.
(That's actually an accidental feature)

The script works in python2 and python3 and only uses the stdlib