configure: bsd sed doesn't support \+ without -r, use * instead
Tested in openbsd 5.6, which does have sed -r (unlike Mac OS X) but it
needs -r to recognize + at all.

The previous commit did fix enough to make it 'work' with the git
describe output as version string, the only problem was that the sed
didn't match in bsds and mac os x.