otr: color multiline messages
Fixes trac ticket 710.

Incoming messages can have newlines in them, which become several
PRIVMSG on the irc layer. Prepending color codes at the beginning of the
message resulted in showing the rest of those PRIVMSG as white.

This splits the message by newlines and rebuilds it in a GString,
re-adding the color codes right after each newline.