The SASL PLAIN RFC says that the first part can be empty
So use the second part as the username, and only require it to be equal
to the first part if that one is present.

ABNF from the spec:

    message   = [authzid] UTF8NUL authcid UTF8NUL passwd

Note brackets.

Authzid (authorization identity) is meant for impersonation, which we
don't support. The actual login username is defined by authcid
(authentication identity)

Thanks grawity for pointing this out.