skype tests: adjust channel length limit to fix tests...
The diff might look confusing - it's just removing the last hex digit of
each channel name in the groupchat tests.

This broke in several ways over the years. I know because i've been
bisecting the last few hours and found lots of different issues.
All of them except this one were fixed already

This one is probably a side effect of the new irc channel name
generation code that includes the # as part of the string and reuses the
nick code. Or maybe not exactly that, but something similar. I'm not
going to bisect any further, had enough for today.