jabber: Fix incoming google talk typing notifications
jabber_buddy_by_jid() does the following:

    head = g_hash_table_lookup(jd->buddies, jid);
    bud = (head && head->next) ? head->next : head;

'head' has the one without resource, 'bud' has the first resource.

So if a resource is available, it uses it and ignores the head.
When asked for a bare JID (with no resource) and GET_BUDDY_EXACT, it
shouldn't do this, but it should return the head.

In other words, the problem was a message in this format:

    <message from="username@gmail.com" ...>

Instead of

    <message from="username@gmail.com/resource" ...>

This only deals with incoming typing notifications. See next commit.