jabber: Fix outgoing google talk typing notifications
Since bare JIDs from typing notifications are now accepted, the other
buddy objects never get the flag JBFLAG_DOES_XEP85 set. This fixes it by
checking both the buddy with resource and the bare one, with the
implication that if the bare JID has that flag, all other resources get
typing notifications.

Follow up to the previous commit, splitting since they are actually
unrelated fixes, although this one is a consequence of the previous one.