yahoo: Fix handling of proxy connection failures
Fixes trac ticket 135, https://bugs.bitlbee.org/bitlbee/ticket/135

The ticket mentions 100% cpu usage when failing to connect to a socks
proxy. It also provides a patch that just checks source == -1 and
disconnects cleanly instead of trying to continue logging in.

This commit is pretty much the same thing, adapted to the API we're
currently using, since the original patch was for bitlbee 1.0.1 and it
looks like the old api was terrible back then.

That ticket was almost 10 years old. The yahoo code didn't change a lot
in that time.