purple: Fix certificate cache dir after changing user dir
This fixes the issue with getting asked to accept certificates that are
perfectly valid, every time.

The directory is normally created by x509_tls_peers_init(), a few calls
below purple_core_init(), which is at module initialization time, way
before we have an irc username to fix the user directory. So it creates
the wrong directory first, and now we have to fix it manually.

And apparently not being able to save cached certificates somehow means
they aren't trusted. For some reason.

< krisfremen> "for some reason"
< dx> idklol