Remove some dust from the old (very old) doc/README
There were some mentions about some password obfuscation method that
probably hasn't been used in a decade. And some portability issues that
are probably not relevant anymore.

Also: "The MD5 algorithm code is licensed under the Aladdin license".
No, it's not. I don't think it was before I removed it in favor of
glib's GChecksum, either.