Fix nicks in non-anonymous jabber groupchats
When I would join a non-anonymous conference room cafe@foo.comm the IRC
nicks were being derived from the participants' jabber ID, and not the
handle with which they joined the room. For example one user in the
roster would be "agarcia" and not "berto". The "berto" name came in the
roster for the room, i.e. a jabber ID of cafe@foo.com/berto.  The name
"berto" doesn't appear anywhere else.  The jabber server is Openfire
3.6.4.

The reason this information wasn't getting propagated correctly appears
to be twofold.  One, the nick_hint wasn't getting set on the buddies for
non-anonymous chats.  Two, setting the nick hint would bail out if the
user was already logged in -- but of course the user is logged in,
that's why I can see them.

Fixes https://bugs.bitlbee.org/bitlbee/ticket/415.