purple: fix /join #channel, which joined a different channel
When joining named channels, purple_chat_join() returned NULL instead of
a struct groupchat, which was actually created in the conversation
created callback (prplcb_conv_new()).

If the name of this channel turned out to be different to the joined
one, this meant having one empty window in the irc client, and another
one for the other channel.

The fix is to return a mostly-empty struct groupchat immediately, and
pick it up later when the PurpleConversation is created using
bee_chat_by_title(). If that fails, fall back to creating a new one.

This bug also meant there was no proper way to report a join failure.
Future commits will address that, this one just makes that easier.

Thanks to Etan Reisner (deryni) for enlightening me while i was trying
to figure out how to fix this.