If imc_away_state_find() fails, try again finding "away"
Fixes trac ticket 1175, https://bugs.bitlbee.org/bitlbee/ticket/1175
"Setting away with libpurple fails silently except for a few values"

Turns out the fallback was the first item of the list, which for
libpurple's jabber is "Feeling" (what?) and seems to be a no-op in
practice.

With this commit the fallbacks are in this order:

1. Try to find whatever the user wrote in the away setting.
2. If that fails, find "away" in the away states.
3. If that fails, use the first away state.

This might not fix it in some rare cases where the prpl doesn't have
an away state equivalent to away. But hopefully the old fallback works
for those, jabber was the weird one here.