purple: fix crash when doing 'chat with' with skypeweb
It was passing the wrong data to the callback - it was supposed to pass
the data of the PurpleMenuItem but it passed the PurpleMenuItem itself.

Probably also applies to other protocols too. It worked fine with
jabber, which i'm guessing is what this code was tested with originally.

It still whines about the null return value saying "(Possible) failure"
but, eh, whatever.