twitter: Changes to 140 char limit detection
As of the 19th of September, Twitter will apparently no longer consider
@usernames at the start of a tweet as part of the 140 characters. URLs are
already ignored when determining this limit, so this patch uses a similar
approach to not counting @usernames.

It's important to note that these are only ignored if at the start of the
tweet, but there's no hard source from Twitter on how this is calculated. For
now, I'm assuming that "@username" at the start is skipping, as is
"@username1 @username2" and so-on.