twitter: fix quoted tweet expansion in extended tweets
The outermost entities object only contains the url of the compat tweet,
the one linking to /i/web/status/[...]

The inner entities object, inside the "extended_tweet", is the one that
contains the quoted tweet url that we're supposed to replace.

But expand_entities() assumed that the quoted_status object would be
next to entities, which doesn't apply in the case of extended tweets.
So now it gets an extra parameter to look for entities.