Update WhatsApp plugin for yowsup 2.5
The removed imports appear to be unnecessary -- nothing else uses those
names in this file -- and they have been removed in yowsup. The other
change is due to a change in API in the YowsupEnv module.

I'm not certain whether this commit is sufficient or even correct, but
it has made the plugin work for me.