purple: hack to pass server parameter to jabber's input request
Because our purple module is all about hacks, adding more can't hurt.

There's a string comparison for "Enter a Conference Server". It's
gettexted in the source but in practice it isn't affected by locale
(bitlbee disables localization).

Worst case, if this stops working, it will open an input request like it
did before this commit. It also does that in purple's jabber if you
don't provide a server parameter.