Include debug symbols in non-debug builds, disable stripping by default
With this commit, the difference between debug and non-debug builds is
mainly the optimization level and -DDEBUG (which isn't used much)

In other words:

    --debug=0 == CFLAGS="-g -O2 -fno-strict-aliasing"
    --debug=1 == CFLAGS="-g3 -DDEBUG -O0"

And --strip=1 can be used to get rid of the debug symbols.

This is closer to the default behavior of autotools.

Should have done this forever ago, like back when bitlbee had bugs (lol)