debian: add preinst script for a smoother transition to systemd units
This fixes an issue when upgrading from an initscripts based release
like 3.4.2 to the current version, in which the old daemon is still
running when systemd tries to start it, so it fails to start due to bind
and fails to stop because it doesn't use pidfiles.

To get around this, the preinst script kills the initscripts-started
daemon the first time when upgrading from something old enough, using
the pidfile.