The protocol argument to socketpair was incorrect.
For some reason the protocol was being set to PF_UNIX which is
incorrect.  According to the man pages and other sources (like Stevens
Unix Network Programming), the protocol should be '0' which will
use the default for the given domain.  This appears to actually
work under Linux (which appears to allow 0 or 1 as the protocol without
error), but fails under Mac OSX (which only allows 0).