twitter: @mention the right people when replying
The previous approach was simply to @ the author of the replied-to
tweet. This is unintuitive
    a) when replying to a tweet that mentions
        several people, in which case they are cut off; and
    b) if you reply to a previous reply you made, in which case you @
        yourself and the intended recipient(s) won't see it.

This commit introduces a function twitter_format_reply to determine
the correct @mention string for the start of your tweet, based on the
original tweet you are replying to. This preserves all mentionees from
the original tweet, removing ourselves if present and making sure
the original tweet's author comes first (if not us).

Mentions are assumed to be space-separated words at the start of the
tweet. Leading and interstitial multiple spaces are handled.

In order to be able to parse the mentions out of the original message,
the text of each tweet is added to the twitter_log_data struct. Replying
to non-logged tweets by tweet ID is still supported, but will use the old
behaviour of just mentioning that tweet's author, whoever it is.