Mastodon: fix pagination issues
When getting the following accounts, we would run into 429 errors. The
spec now reads: "Note: max_id and since_id for next and previous pages
are provided in the Link header. It is not possible to use the id of
the returned objects to construct your own URLs, because the results
are sorted by an internal key."
https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#getting-who-account-is-following

Also handle delete notifications.