Mastodon: get rid of the md->http_fails counter
I'm not sure why this was part of the Twitter API. I find it annoying
because I must repeat commands multiple times before seeing a decent
error, eg. when favoring the same toot multiple times, all attempts
except for the first one are errors, but I can only see this when
debugging. This is not good.