Mastodon: whois prints handle
When using the REPLY <id> <text> command, your text should mention the
various handles in the original message beyond the original author you
are replying to (that one is already added). All the user has
available, however, is nicks. The WHOIS command allows the user to
figure out that kensanata_ is actually @kensanata and that kensanata__
is actually @kensanata@mammouth.cafe. In addition to that, if you
provide the id of a status, it fetches that status from the server and
prints all the mentions therein.