Mastodon: fix replies to self
When replying to self, we need a bu associated with the original
status or bu->handle will bomb. This commit adds such a handle to the
log of any status of ours we display such that it can be retrieved
later when replying to it.

This commit also suppresses our own account name when replying to
ourselves.