Mastodon 2.0 support: ids are strings
"All id attributes in the REST API responses, including attributes
that end in _id, are now returned as strings instead of integers. This
is because large integers cannot be encoded in JSON losslessly, and
all IDs in Mastodon are now bigint."
https://medium.com/@Gargron/mastodon-2-0-e93d9d28dbb9

I'm assuming this refers to Ruby on Rails, where the bigint uses 8
bytes, ie. 64 bits.
http://mytrile.github.io/blog/2013/05/16/integer-types-in-rails/

In C using glib, a guint64 also 64bits but unsigned.
https://developer.gnome.org/glib/stable/glib-Basic-Types.html#guint64

This patch handles ids as strings, if and only if they are not
negative.