irc_commands: send numeric 900 RPL_SASLSUCCESS too, some clients need it
The nodejs irc-framework lib used by thelounge and kiwiirc seems to
expect 900 on successful login, instead of just 903, so they hang on
sasl auth from bitlbee.

The sasl spec isn't very specific on which should be used (IMO clients
should handle both), but it seems to point at the direction of sending
both, so both it is.