debian: turn -common and -dev into arch:any to safely use --link-doc
This started with trying to remove something that says "remove the hack
below once Hardy and Lenny are deprecated". Hardy and Lenny have been
deprecated for years, this should be easy, right?

Turns out that the symlink stuff we were doing can result in broken
packages with binNMU, see debbug#767839. This is also documented in the
dh_installdocs man page:

> CAVEAT 2: The use of --link-doc should only be done when the packages
> have same "architecture" type. A link from a architecture independent
> package to a architecture dependent package (or vice versa) will not
> work. Since compat 10, debhelper will actively reject unsupported
> combinations.

And we were doing essentially the same wrong thing but manually.

The easiest way to solve this is to stop mixing arch:any and arch:all,
and to make everything arch:any. Let's hope this is as easy as it looks.

Thanks to bunk, nthykier and olly from #debian-devel for help with this.