twitter: show quote tweets when the url isn't in the entities
Which means we recognized that a tweet had a quote in it, but not
finding a matching URL entity meant we didn't have anywhere to show it.
Now it's appended at the end.

This is what changed here: https://twittercommunity.com/t/105473
("Updating how URLs are rendered in the Quote Tweet payload")

Took me way too long to realize it was this and why it happened - if you
create quote tweets by copypasting urls, the behavior doesn't change,
this only applies if you use the retweet button in twitter web and add a
comment.

And initially we disregarded it as some brownout of the userstream, but
it looks like it applies to the REST api too.