Add support for using libmd for MD5 functions
libmd is a library found on some Linux distributions that provides MD5
and other message digest functions.  It is generally found as a
companion to the libbsd library.

This patch adds --external_md5=0/1/auto to let the builder select
whether or not to build with the internal MD5 routines or link with
libmd.

Signed-off-by: David Cantrell <dcantrell@redhat.com>