Handle HTTP 308 the same as 307
This makes the HTTP client code handle status code 308 (Permanent
Redirect) in a similar manner to the already existing support for 307
(Temporary redirect). Like 307, no change of method is applied to the
redirect.

The use case for this has arisen from the bitlbee-mastodon plugin, the
Mastodon API returns 308 for some requests.